requirements.txt
line-bot-sdk
bs4
flask
datetime
pandas
emoji
scipy
gunicorn